% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations.R
\name{add_observations_data}
\alias{add_observations_data}
\title{Add observations data to data frame}
\usage{
add_observations_data(data, which, by = "observations", dataset)
}
\arguments{
\item{data}{data frame}

\item{which}{observations data columns to add}

\item{by}{column of the observations data to join on (default "observations")}

\item{dataset}{dataset}
}
\value{
the data frame with the requested observations data columns added
}
\description{
Add observations data to data frame
}
