% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables.R
\name{add_variables_data}
\alias{add_variables_data}
\title{Add variables data to data frame}
\usage{
add_variables_data(data, which, by = "variables", dataset)
}
\arguments{
\item{data}{data frame}

\item{which}{variables data columns to add}

\item{by}{column of the variables data to join on (default "variables")}

\item{dataset}{dataset}
}
\value{
the data frame with the requested variables data columns added
}
\description{
Add variables data to data frame
}
