% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{add_y}
\alias{add_y}
\title{Add y-vlaues to a ggplot}
\usage{
add_y(position = "y", size = 3, vjust = -0.6, dodge = 0.75)
}
\arguments{
\item{position}{where to place the label ("max", "min" or "mean")}

\item{size}{text size of the label}

\item{vjust}{vertical justification of the label}

\item{dodge}{width passed to \code{position_dodge()}, to match a dodged geom}
}
\description{
Add y-vlaues to a ggplot
}
