% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{alignexp_tibble_rows}
\alias{alignexp_tibble_rows}
\title{Align and expand rows of a list of tibbles}
\usage{
alignexp_tibble_rows(x, by = NULL)
}
\arguments{
\item{x}{list of tibbles}

\item{by}{column to join and align}
}
\description{
Chooses the first column if nothing is supplied to argument by.
}
