% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_character_null}
\alias{as_character_null}
\title{Convert Variable to character including NULL}
\usage{
as_character_null(x)
}
\arguments{
\item{x}{An object to coerce to character}
}
\value{
returns a character vector
}
\description{
Convert Variable to character including NULL
}
