% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{button_begin}
\alias{button_begin}
\title{Bootstrap button to hide/show segments}
\usage{
button_begin(label = "View/Hide")
}
\arguments{
\item{label}{text shown on the button}
}
\description{
Bootstrap button to hide/show segments
}
