% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{button_end}
\alias{button_end}
\title{Bootstrap button to hide/show segments - simple div to end button}
\usage{
button_end()
}
\description{
Bootstrap button to hide/show segments - simple div to end button
}
