% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ordering.R
\name{cap_n}
\alias{cap_n}
\title{Convert number vector to character and summarize x >= y to "y+"}
\usage{
cap_n(x, y = 5)
}
\arguments{
\item{x}{vector of numbers}

\item{y}{maximum number of characters to keep}
}
\description{
Convert number vector to character and summarize x >= y to "y+"
}
