% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catOSWruns.R
\name{catOSWruns_}
\alias{catOSWruns_}
\title{Concatenate different run OSW tables into one table}
\usage{
catOSWruns_(
  sqMass_files,
  in_osw,
  which_m_score = "m_score",
  m_score_filter = 0.05,
  report_top_single_result = T,
  run_sub_expression = NULL,
  mod_exclusion = NULL,
  mod_grouping = NULL,
  ...
)
}
\arguments{
\item{sqMass_files}{An array of character paths to the chromatograms that have corresponding names in the OSW file.}

\item{in_osw}{A character vector of the absolute path and filename of the osw file. (Must be .osw format)}

\item{which_m_score}{A character vector indicating which m_score to use. (Options: m_score or ms2-m_score. Default: 'm_score')}

\item{m_score_filter}{A numeric vector indicating the threshold m_score cut-off value for filtering. (Defualt: 0.05)}

\item{report_top_single_result}{A logical value indificating to keep only the results with the lowest m_score. (Default: TRUE)}

\item{mod_exclusion}{A vector of unimod record id modifications to exclude, and replace by nothing. Example: c("35")}

\item{mod_grouping}{A dataframe contained uninod ids and random characters to group isomers by. Example: data.frame( record_id=c(4, 259, 267), rand_char_id="B", stringsAsFactors = F )}
}
\value{
A data.table containing Run ID information
}
\description{
This function can be used to extract OSW data for several runs from an OSW file, and concatenate
them into one table with a run id
}
\author{
Justin Sing \url{https://github.com/singjc}
}
