% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cc}
\alias{cc}
\title{Vector with elements as names}
\usage{
cc(...)
}
\arguments{
\item{...}{vector/s whose elements become their own names}
}
\value{
a character vector named by its own elements
}
\description{
Vector with elements as names
}
