% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sqlite_table}
\alias{check_sqlite_table}
\title{check_sqlite_table}
\usage{
check_sqlite_table(conn, table, msg = "")
}
\arguments{
\item{conn}{Connection to database}

\item{table}{Character vector to test for, if present in database}

\item{msg}{A character to pre-append to stop error message. (Optional)}
}
\value{
Logical value, TRUE if table is present
}
\description{
check_sqlite_table
}
