% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ordering.R
\name{classify_0NANaNInf}
\alias{classify_0NANaNInf}
\title{Returns character label for various types of vecrtor elements}
\usage{
classify_0NANaNInf(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Returns character label for various types of vecrtor elements
}
