% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnimodMapping.R
\name{codenameTounimod}
\alias{codenameTounimod}
\title{Conversion from modification codename to UniMod ID}
\usage{
codenameTounimod( mod_seq, out='sequence' )
}
\arguments{
\item{mod_seq}{A character vector containing of a peptide sequence, containing modification code names to be converted to UniMod IDs}
}
\description{
Conversion from modification codename to UniMod ID
}
\examples{
mod_seq <- "EGHAQNPM(Oxidation)EPS(Phospho)VPQLS(Phospho)LM(Oxidation)DVK"
## Will return peptide sequence with UniMod ID
codenameTounimod ( mod_seq )
## Out
## [1] "EGHAQNPM(UniMod:35)EPS(UniMod:21)VPQLS(UniMod:21)LM(UniMod:35)DVK"
}
