% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2volume.R
\name{counts2volume}
\alias{counts2volume}
\title{counts2volume}
\usage{
counts2volume(input, scaling_factor, fractions = FALSE)
}
\arguments{
\item{input}{Data frame. First column should be diameter bins. The remaining columns will be treated as counts to be transformed.}

\item{scaling_factor}{Factor used to estimate volume from diameters. n in 4/3*pi*r^n.}

\item{fractions}{Return volumetric fraction within each bin rather than volume per bin}
}
\description{
This function translates counts total volume (rather than counts) contained within each diameter bin.
}
\examples{
counts2volume(input, scaling_factor)
}
\keyword{multisizer}
