% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSWReportsVisualization.R
\name{d_score_hist}
\alias{d_score_hist}
\title{Plot distribution of d-score}
\usage{
osw_reports_visualization$d_score_hist( osw, bins=20, position= 'dodge2' )
}
\arguments{
\item{osw}{A data.table/data.frame of the OSW results}

\item{bins}{A numeric vector for number of bins to bin data into}

\item{position}{A character vector indicating how groupings should be plot. (Default: 'dodge2') See ggplot2:geom_histogram}
}
\description{
Plot distribution of d-score
}
