% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{data_frame2tibble}
\alias{data_frame2tibble}
\title{Transforms data frames to tibble and adds column for row names}
\usage{
data_frame2tibble(data, to.row.names = "rows")
}
\arguments{
\item{data}{data frame with row names}

\item{to.row.names}{name for row names column}
}
\description{
Transforms data frames to tibble and adds column for row names
}
