% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_diann.R
\name{.add_defaults}
\alias{.add_defaults}
\title{Registers the default column sets used by the DIA-NN importers}
\usage{
.add_defaults()
}
\value{
nothing; \code{Info$defaults} is created in the global environment
}
\description{
Registers the default column sets used by the DIA-NN importers
}
