% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{.axis_limit_breaks}
\alias{.axis_limit_breaks}
\title{Helps make nice axis limit breaks}
\usage{
.axis_limit_breaks(plot.limits, break.size = NULL)
}
\arguments{
\item{plot.limits}{vector containing plot limits}

\item{break.size}{space between breaks}
}
\value{
list with elements \code{limits} and \code{breaks}
}
\description{
Helps make nice axis limit breaks
}
