% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_data.R
\name{.cat_function}
\alias{.cat_function}
\title{Prints function snippet to console and clipboard}
\usage{
.cat_function(FUN, add.equal.sign = T)
}
\arguments{
\item{FUN}{function}

\item{add.equal.sign}{add equal sign to add arguments}
}
\value{
FALSE if \if{html}{\out{<FUN>}} is not a function, otherwise the snippet is printed
}
\description{
Prints function snippet to console and clipboard
}
