% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_diann.R
\name{.get_defaults}
\alias{.get_defaults}
\title{Returns a default column set used by the DIA-NN importers}
\usage{
.get_defaults(type, name)
}
\arguments{
\item{type}{default set to return ("variables_data" or "data_frames")}

\item{name}{name of a single default within \if{html}{\out{<type>}}}
}
\value{
the requested defaults, or the whole list if \if{html}{\out{<type>}} is missing
}
\description{
Returns a default column set used by the DIA-NN importers
}
