% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{.get_plot_limits}
\alias{.get_plot_limits}
\title{Extract the built x/y axis ranges of a ggplot}
\usage{
.get_plot_limits(plot)
}
\arguments{
\item{plot}{a ggplot object}
}
\value{
list with \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}
}
\description{
Extract the built x/y axis ranges of a ggplot
}
