% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateVolumeFactors.R
\name{estimateVolumeFactors}
\alias{estimateVolumeFactors}
\title{estimateVolumeFactors}
\usage{
estimateVolumeFactors(input, real_vol, start = 2, end = 3,
  increment = 0.01, df = FALSE)
}
\arguments{
\item{input}{Data frame with two columns, the first with diameter bins, the second with counts per ml.}

\item{start}{Low end to begin searching for optimal factor}

\item{end}{High end of range to search for optimal factor}

\item{increment}{Amount to increment when conducting search for optimal factor}

\item{df}{Return data frame of volume factor estimates and differences from actual volume instead of best fit volume factor}

\item{total_vol}{Known total volume per ml in sample. Can be obtained from sonicated sample measurments.}
}
\description{
This function returns the estimated value of n in 4/3*pi*r^n that minimizes the difference between calculated total volume and real total volume (as estimated via sonicated samples).
}
\examples{
estimateVolumeFactors(input, total_vol)
}
\keyword{multisizer}
