% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_sizecounts.R
\name{expand_sizecounts}
\alias{expand_sizecounts}
\title{expand_sizecounts}
\usage{
expand_sizecounts(counts, vol_col = 1, diam_col = 2)
}
\arguments{
\item{counts}{Data frame of binned counts per size bin. First two columns should correspond to volume and diameter bins. The final columns should correspond to sample counts.}

\item{vol_col}{Column in data frame containing volume data}

\item{diam_col}{Column in data frame containing diameter data}
}
\description{
This function expands counts into a format that allows you to plot distributions.
}
\examples{
expand_sizecounts(counts)
}
\keyword{multisizer}
