% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendDiameterRange.R
\name{extendDiameterRange}
\alias{extendDiameterRange}
\title{extendDiameterRange}
\usage{
extendDiameterRange(main_df, short_df)
}
\arguments{
\item{main_df}{Data frame of counts containing extended range of diameter values}

\item{short_df}{Data frame you wish to add extended range rows to}
}
\description{
This function adds extra rows to a data frame not containing the multisizers extended diameter range.
}
\examples{
extendDiameterRange(main_df, short_df)
}
\keyword{multisizer}
