% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOSWdb.R
\name{filterOSWdb}
\alias{filterOSWdb}
\title{Filter and OSW db file}
\usage{
filterOSWdb(
  osw_file,
  unmodified_sequence_filter = NULL,
  modified_sequence_filter = NULL,
  random_seed = NULL,
  ratio_keep = NULL
)
}
\arguments{
\item{osw_file}{A character vector of the absolute path and filename of the osw file. (Must be .osw format)}

\item{unmodified_sequence_filter}{A character vector for extraction of specific peptide(s). I.e. c('ANSSPTTNIDHLK', 'ESTAEPDSLSR', 'NLSPTKQNGKATHPR', 'KDSNTNIVLLK', 'NKESPTKAIVR')}

\item{modified_sequence_filter}{A character vector for extraction of specific peptide(s) with modifications}

\item{random_seed}{(numeric) Set the seed for sampling}

\item{ratio_keep}{(float) Set the fraction of sequences to keep}
}
\value{
A data.table containing spectral library information
}
\description{
This function can be used to filter an osw db file given a list of unmodified sequences
}
\author{
Justin Sing \url{https://github.com/singjc}
}
