% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPQPdb.R
\name{filterPQPdb}
\alias{filterPQPdb}
\title{Filter and PQP db file}
\usage{
filterPQPdb(pqp_file, unmodified_sequence_filter)
}
\arguments{
\item{pqp_file}{A character vector of the absolute path and filename of the sqMass file. (Must be .osw format)}

\item{unmodified_sequence_filter}{A character vector for extraction of specific peptide(s). I.e. c('ANSSPTTNIDHLK', 'ESTAEPDSLSR', 'NLSPTKQNGKATHPR', 'KDSNTNIVLLK', 'NKESPTKAIVR')}
}
\value{
A data.table containing spectral library information
}
\description{
This function can be used to filter an PQP db file given a list of unmodified sequences
}
\author{
Justin Sing \url{https://github.com/singjc}
}
