% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnimodMapping.R
\name{generateUniModMapping_}
\alias{generateUniModMapping_}
\title{Generate a UniMod mapping table}
\usage{
generateUniModMapping_(unimod_xml = NULL)
}
\arguments{
\item{unimod_xml}{A character path indicating the path to a unimod.xml path}
}
\value{
A data.table containing UniMod ID to Modification Code Name mapping
}
\description{
This function can be used to generate a data.table containing
UniMod ID to Modification code name mapping
}
\author{
Justin Sing \url{https://github.com/singjc}
}
