% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations.R
\name{get_observations}
\alias{get_observations}
\title{Return observations}
\usage{
get_observations(observations, dataset, ...)
}
\arguments{
\item{observations}{vector of observations}

\item{dataset}{dataset}

\item{...}{additional arguments}
}
\value{
a vector of observation identifiers
}
\description{
Return observations
}
