% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFDREstimatePerformance.R
\name{getFDREstimatePerformance}
\alias{getFDREstimatePerformance}
\title{get qunatification per run plot}
\usage{
getFDREstimatePerformance(
  data,
  p.title = NULL,
  true_label = "TP",
  global_fdr_label = "ipf_fdr",
  local_fdr_label = "ipf_lfdr"
)
}
\arguments{
\item{data}{A data.table containing intensity information}

\item{p.title}{A character to label plot}

\item{true_label}{A character vector indicating the column header name containing true class labels}

\item{global_fdr_label}{A character vector indicating the column header name containing the global fdr estimates}

\item{local_fdr_label}{A chracter vector indicating the column header name containing the local fdr estimates}
}
\value{
A list containg a data.table of the data used to make performance and plot, and a ggplot object of the performance plot
}
\description{
This function can be used to generate a qunatification box plot per run
}
\author{
Justin Sing \url{https://github.com/singjc}
}
