% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdentificationsBarplot.R
\name{getIdentificationsBarplot}
\alias{getIdentificationsBarplot}
\title{get qunatification per run plot}
\usage{
getIdentificationsBarplot(
  data,
  data_to_show = NULL,
  include_GT = NULL,
  var_order = NULL,
  show_cumulative = FALSE,
  fdr_threshold = 0.05,
  p.title = NULL,
  dataset = ""
)
}
\arguments{
\item{fdr_threshold}{A numeric vector to add fdr threshold lines to the plot. Default: fdr_threshold =  0.05}

\item{p.title}{A character to label plot}

\item{dataset}{A character vector specifying special label mappings for specific dataset. Options: "", "Synth_Dilution"}

\item{osw_df_bm}{A data.table containing osw results with bitmap information (TPs, FPs, TNs, FNs)}

\item{ipf_df_bm}{A data.table containing ipf results with bitmap information}
}
\value{
A ggplot object
}
\description{
This function can be used to generate a qunatification box plot per run
}
\author{
Justin Sing \url{https://github.com/singjc}
}
