% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModificationPosition.R
\name{getModificationPosition_}
\alias{getModificationPosition_}
\title{Get Modificiation position from modified sequence. Currently only support phosphorylation.}
\usage{
getModificationPosition_(mod_seq, character_index = F)
}
\arguments{
\item{mod_seq}{A character vector of the modified position.}

\item{character_index}{A Logical value indicating to return the exact location of ammino acid modified, or the index of modified text in string.}
}
\value{
A numeric value for modification position
}
\description{
This function can be used to get the modification position for a given modified sequence
Currently only support phosphorylation/UniMod:21. TODO: Make it useable for other cases
}
\author{
Justin Sing \url{https://github.com/singjc}
}
