% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzPntrs.R
\name{getmzPntrs}
\alias{getmzPntrs}
\title{Get a list of mz Objects}
\usage{
getmzPntrs(dataPath, runs)
}
\arguments{
\item{dataPath}{data path that contains mzml folder of chromtogram files}

\item{runs}{A list variable containing paths to chromatogram files}
}
\value{
(A list of mzRpwiz)
}
\description{
Get a list of mz Objects
}
