% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSWData.R
\name{getOSWData_}
\alias{getOSWData_}
\title{Extract data from OpenSwathWorkflow .osw results file}
\usage{
getOSWData_(oswfile, run_name = "", precursor_id = "",
  peptide_id = "", mod_peptide_id = "", mod_residue_position = "",
  peak_group_rank_filter = FALSE, pep_list = "", mscore_filter = "",
  ipf_filter = "", ipf_score = FALSE, ms2_score = TRUE,
  decoy_filter = TRUE)
}
\arguments{
\item{oswfile}{A character vector of the absolute path and filename of the osw results file. (Must be .osw format)}

\item{run_name}{A character vector for extraction of a specific run, this should be the same as the file name in the .OSW RUN table. (i.e. run_name='yanliu_I170114_016_PhosNoco4_SW.mzXML.gz', Default: '')}

\item{precursor_id}{A numeric value for a specific precursor id to extract information for. (Default: '')}

\item{peptide_id}{A string vector indicating a specific peptide to extract information for. (Default: '')}

\item{mod_peptide_id}{An array of two string vectors indicating a specific modified peptide sequence with both UniMod annotation and actual modification name to extract information for. I.e. c(ANS(Phos)SNSLK, ANS(UniMod:21)SNSLK) (Default: '')}

\item{mod_residue_position}{A numeric value indicating the position of a modification. (Default: '')}

\item{peak_group_rank_filter}{A Logical value for filtering OSW results by Peak-Group Rank of 1. (Default: FALSE)}

\item{pep_list}{An arrary of string modified peptide sequences to extract information for. (Default: '')}

\item{mscore_filter}{A numeric value to filter results by MS2 q-value. (Default: '')}

\item{ipf_filter}{A numeric value to filter results by IPF PEP. (Default: '')}

\item{ipf_score}{A logical value to extract data using IPF Score results. (Default: FALSE)}

\item{ms2_score}{A logical value to extract data using MS2 Score results. (Default: TRUE)}

\item{decoy_filter}{A logical value to filter decoys out of final results. (Default: TRUE)}
}
\value{
A data.table containing OpenSwath Results information.
}
\description{
This function can be used to extract information from the OSW results file obtained from
running OpenSwathWorkflow
}
\author{
Justin Sing \url{https://github.com/singjc}
}
