% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPepLibData.R
\name{getPepLibData_}
\alias{getPepLibData_}
\title{Extract data from a PQP library file}
\usage{
getPepLibData_(libfile, peptide_id = "", mod_peptide_id = c("", ""))
}
\arguments{
\item{libfile}{A character vector of the absolute path and filename of the library file. (Must be .pqp format)}

\item{peptide_id}{A character vector for extraction of a specific peptide. I.e. 'ANSSPTTNIDHLK'}

\item{mod_peptide_id}{An array of two string vectors indicating a specific modified peptide sequence with both UniMod annotation and actual modification name to extract information for. I.e. c(ANS(Phos)SNSLK, ANS(UniMod:21)SNSLK) (Default: '')}
}
\value{
A data.table containing spectral library information
}
\description{
This function can be used to extract information from the PQP library file. This can also be used to extract the transition informaiton from and OSW results file.
}
\author{
Justin Sing \url{https://github.com/singjc}
}
