% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuantificationplot.R
\name{getQuantificationplot}
\alias{getQuantificationplot}
\title{get qunatification per run plot}
\usage{
getQuantificationplot(
  data,
  fdr_threshold = 0.05,
  scores = "ipf_fdr",
  intensities = "ipf_Intensity",
  true_peptides = "ipf_TP",
  normalzied_relative_intensity = FALSE,
  keep.only.TP = TRUE,
  p.title = NULL,
  dataset = ""
)
}
\arguments{
\item{data}{A data.table containing intensity information}

\item{fdr_threshold}{A numeric vector to add fdr threshold lines to the plot. Default: fdr_threshold =  0.05}

\item{scores}{A character vector of the data column name scores to filter data on}

\item{intensities}{A character vector inficating the column to use for intentisty information}

\item{true_peptides}{A character vector indicating the column header name containing true class labels}

\item{keep.only.TP}{A logical value indicating to only show intensity information for true positive peptides}

\item{p.title}{A character to label plot}
}
\value{
A  a ggplot object
}
\description{
This function can be used to generate a qunatification box plot per run
}
\author{
Justin Sing \url{https://github.com/singjc}
}
