% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getROC.R
\name{getROC}
\alias{getROC}
\title{get Receiver Operator Characterostocs plot and data}
\usage{
getROC(data, scores, p.title = NULL, fdr_threshold = c(0.01, 0.05))
}
\arguments{
\item{data}{A data.table containing p-value/q-value score and true class labels}

\item{scores}{A character vector of the data column name scores to compute the pseudo-recall for}

\item{p.title}{A character to label plot}

\item{fdr_threshold}{A numeric vector to add fdr threshold lines to the plot. Default: fdr_threshold = c(0.01, 0.05)}
}
\value{
A list containing the ROC data and  a ggplot
}
\description{
This function can be used to generate the pseudo-recall plot
}
\author{
Justin Sing \url{https://github.com/singjc}
}
