% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunID.R
\name{getRunID_}
\alias{getRunID_}
\title{Extract Run ID information from a osw file fiven a run name}
\usage{
getRunID_(oswfile, run_name)
}
\arguments{
\item{oswfile}{A character vector of the absolute path and filename of the osw file. (Must be .osw format)}

\item{run_name}{A character vector for extraction of a specific run.}
}
\value{
A data.table containing Run ID information
}
\description{
This function can be used to extract the Run ID for a given run
}
\author{
Justin Sing \url{https://github.com/singjc}
}
