% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScalingFactors.R
\name{getScalingFactors}
\alias{getScalingFactors}
\title{getScalingFactors}
\usage{
getScalingFactors(df, sonication_indicator = ".S")
}
\arguments{
\item{df}{Data frame of binned counts per size bin.}

\item{sonication_indicator}{Character string in sample names used to designate component parts of whole}
}
\description{
Returns scaling factors for data frame containing whole counts and corresponding component (e.g. sonicated) counts.
}
\examples{
getScalingFactors(df)
}
\keyword{multisizer}
