% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsqMassPntrs.R
\name{getsqMassPntrs}
\alias{getsqMassPntrs}
\title{Get a list of data.tables for chromatrogram id mappings}
\usage{
getsqMassPntrs(
  dataPath,
  runs,
  nameCutPattern = "(.*)(/)(.*)",
  chrom_ext = ".chrom.sqMass"
)
}
\arguments{
\item{input}{A shiny input variable that contains Working Directory Information}

\item{global}{A list variable containing paths to chromatogram files}
}
\value{
(A list of mzRpwiz)
}
\description{
Get a list of data.tables for chromatrogram id mappings
}
