% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_diann.R
\name{import_diann}
\alias{import_diann}
\title{Imports a DIA-NN report into a dataset store}
\usage{
import_diann(
  file = "report.parquet",
  name = "Precursors",
  filter_by = Proteotypic == 1 & Decoy == 0,
  observation_names = list(pattern = ".+"),
  variables_data = "default",
  data_frames = "default",
  preview_precursors = c("all", "top100"),
  preview_format = c("wide_obs", "wide_vars", "long"),
  save_dir,
  partition_by_run = F,
  silent = F
)
}
\arguments{
\item{file}{path to a DIA-NN report (parquet or tsv)}

\item{name}{name of the dataset to create}

\item{filter_by}{expression used to filter the report rows}

\item{observation_names}{list describing how run names become observations}

\item{variables_data}{name of a default column set, or column names}

\item{data_frames}{name of a default quantity set, or column names}

\item{preview_precursors}{how many precursors to keep in the preview}

\item{preview_format}{shape of the preview ("wide_obs", "wide_vars", "long")}

\item{save_dir}{folder the dataset store is written to}

\item{partition_by_run}{partition the saved data frames by run}

\item{silent}{Should messages be suppressed?}
}
\value{
TRUE (invisibly); the dataset is written to <save_dir> and
registered in Datasets/.Datasets
}
\description{
Imports a DIA-NN report into a dataset store
}
