% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listTostring}
\alias{listTostring}
\title{Convert list object to printable character vectory}
\usage{
listTostring(list_obj, collapse = "\\n")
}
\arguments{
\item{list_obj}{A list object to coerce to character}

\item{collapse}{A character vector to collapse characters on}
}
\value{
returns a character vector
}
\description{
Convert list object to printable character vectory
}
