% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_setup.R
\name{log_setup}
\alias{log_setup}
\title{Logging system setup}
\usage{
log_setup()
}
\description{
Setup logging system to write out error, trace, warnings and general information to file
}
\author{
Justin Sing \url{https://github.com/singjc}
}
