% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstools-package.R
\docType{package}
\name{mstools-package}
\alias{mstools}
\alias{mstools-package}
\title{mstools: Tool functions that can be used by several madrat-dependent or magpie4 output functions}
\description{
Tool functions that can be used by several madrat-dependent or magpie4 output functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/magpie4}
  \item \doi{10.5281/zenodo.1158582}
  \item Report bugs at \url{https://github.com/pik-piam/magpie4/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Leon Bodirsky \email{bodirsky@pik-potsdam.de}

Authors:
\itemize{
  \item Benjamin Leon Bodirsky \email{bodirsky@pik-potsdam.de}
  \item Kristine Karstens
  \item Felicitas Beier
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}
  \item Pascal Sauer \email{pascal.sauer@pik-potsdam.de} (\href{https://orcid.org/0000-0002-6856-8239}{ORCID}) (affiliation: Potsdam Institute for Climate Impact Research)
}

}
