% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00d_data_frames.R
\name{open_data_frame}
\alias{open_data_frame}
\title{Assemble data frame from dataset}
\usage{
open_data_frame(which, variables, observations, dataset, ...)
}
\arguments{
\item{which}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{dataset}{dataset name}

\item{...}{additional arguments for open_data()}
}
\value{
a lazy Arrow connection to the assembled data frame
}
\description{
Assemble data frame from dataset
}
