% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSWReportsVisualization.R
\docType{data}
\name{osw_reports_visualization}
\alias{osw_reports_visualization}
\title{mstools osw reports visualization environment for various stats plotting functions}
\format{An object of class \code{environment} of length 1.}
\usage{
osw_reports_visualization
}
\description{
This function encapsulates various plotting functions that can be used for plotting
various reports from the OSW file.
}
\author{
Justin Sing \url{https://github.com/singjc}
}
\keyword{datasets}
