% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparison_vectorized.R
\name{pairwise_comparison_vectorized}
\alias{pairwise_comparison_vectorized}
\title{Pair-Wise Comparison vectorized}
\usage{
pairwise_comparison_vectorized(
  leftWidth.x,
  leftWidth.y,
  rightWidth.x,
  rightWidth.y,
  out = "samePeak",
  P_overlap_threshold = 0.2
)
}
\arguments{
\item{leftWidth.x}{left border of comparison 1}

\item{leftWidth.y}{left border of comparison 2}

\item{rightWidth.x}{right border of comparison 1}

\item{rightWidth.y}{right border of comparison 2}

\item{out}{a character vector specifying method to perform. Options: ('delta', 'strcat', 'isequal', 'samePeak', 'percentOverlap')}

\item{P_overlap_threshold}{a numeric value specifying the curret off to use for considering a peak overlap as bein the same or not}
}
\value{
A logical value if boundaries are comparable
}
\description{
This function can be used to perform a pairwise comparison for vectors of boundaries
}
\author{
Justin Sing \url{https://github.com/singjc}
}
