% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.R
\name{parseMS}
\alias{parseMS}
\title{Parse output from MS}
\usage{
parseMS(file)
}
\arguments{
\item{file}{filename to MS simulation results.}
}
\value{
A list containing each simulation's data.
}
\description{
\code{parseMS} parses results from an MS simulation, returning a list of results.
}
\examples{
res <- parseMS(system.file("extdata", "ms-01.sim", package="mstools"))
summary(sapply(res, function(x) x$segsites))
}

