% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ordering.R
\name{partition_n}
\alias{partition_n}
\title{Transform values of x (row numbers) into groups of n}
\usage{
partition_n(x, n)
}
\arguments{
\item{x}{vector to work on}

\item{n}{number of partitions}
}
\description{
Transform values of x (row numbers) into groups of n
}
