% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pdf_open}
\alias{pdf_open}
\title{Title}
\usage{
pdf_open(file)
}
\arguments{
\item{file}{pdf file to open (opens file saved under .pdf_temp if no argument given)}
}
\description{
Title
}
\examples{
pdf_temp()
plot(6, 9)
pdf_open
}
