% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pdf_temp}
\alias{pdf_temp}
\title{Title}
\usage{
pdf_temp(width = 5, height = 5, ...)
}
\arguments{
\item{width}{the width of the graphics region in inches (default = 5)}

\item{height}{the height of the graphics region in inches (default = 5)}

\item{...}{additional parameters for pdf()}
}
\description{
Title
}
\examples{
pdf_temp()
plot(6, 9)
pdf_open
}
