% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_ggtext}
\alias{plot_ggtext}
\title{Plots text on an empty ggplots}
\usage{
plot_ggtext(text = "Lorem ipsum whatever", size = 8)
}
\arguments{
\item{text}{text to display}

\item{size}{text size}
}
\description{
Plots text on an empty ggplots
}
