% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve.R
\name{plot_standard_curves}
\alias{plot_standard_curves}
\title{Plot Standard Curves for LC-MS Analysis}
\usage{
plot_standard_curves(data, compounds = "all", log = FALSE, split = FALSE)
}
\arguments{
\item{data}{A data frame containing concentration and peak area data from LC-MS analysis}

\item{compounds}{Character vector of compound names or "all" for all compounds}

\item{log}{Logical, whether to use log transformation for better visualization}

\item{split}{Logical, whether to save plots separately for each compound}
}
\value{
A list of linear regression results for each compound
}
\description{
This function creates standard curves for LC-MS compounds analysis. It supports
both individual and combined plots with options for log transformation.
}
\examples{
\dontrun{
# Create example data from LC-MS analysis
example_data <- data.frame(
  Conc = c(5, 2.5, 1.25, 0.625, 0.3125, 0.15625, 0.078125),
  AA = c(68619120, 29206615, 16247479, 7586211, 4235692, 1846262, 1121494),
  DHA = c(58619120, 25206615, 14247479, 6586211, 3235692, 1646262, 921494),
  EPA = c(48619120, 22206615, 12247479, 5586211, 2935692, 1446262, 821494),
  ALA = c(38619120, 19206615, 10247479, 4586211, 2535692, 1246262, 721494),
  LA = c(28619120, 16206615, 8247479, 3586211, 2135692, 1046262, 621494),
  GLA = c(18619120, 13206615, 6247479, 2586211, 1735692, 846262, 521494)
)
# Plot standard curves for LC-MS compounds
plot_standard_curves(example_data, compounds = c("AA", "DHA"), log = TRUE, split = FALSE)
}
}
