% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{plot_tabset}
\alias{plot_tabset}
\title{Plot the markdown code for individual tabsets}
\usage{
plot_tabset(.x, .pf, header = c("## ", " {-}"), options = "warning=FALSE")
}
\arguments{
\item{.x}{verctor to iterate through for each tabset panel}

\item{.pf}{plotting function}

\item{header}{text to add before and after .x as header}

\item{options}{options to add before plotting function}
}
\description{
Plot the markdown code for individual tabsets
}
