% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{png_open}
\alias{png_open}
\title{Title}
\usage{
png_open(file)
}
\arguments{
\item{file}{png file to open (opens file saved under .png_temp if no argument given)}
}
\description{
Title
}
\examples{
png_temp()
plot(6, 9)
png_open
}
