% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{png_temp}
\alias{png_temp}
\title{Title}
\usage{
png_temp(..., width = 10, height = 5, units = "cm", res = 300)
}
\arguments{
\item{...}{additional parameters for png()}

\item{width}{the width of the graphics region in the given unit
(default = 10)}

\item{height}{the height of the graphics region in the given unit
(default = 5)}

\item{units}{The units in which height and width are given. Can be cm
(default), mm, px (pixels) or in (inches)}

\item{res}{the resolution in ppi}
}
\description{
Title
}
\examples{
png_temp()
plot(6, 9)
png_open
}
