% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_msdial_to_jupyter.R
\name{run_msdial_to_jupyter}
\alias{run_msdial_to_jupyter}
\title{MSdial to Jupyter Converter}
\usage{
run_msdial_to_jupyter()
}
\value{
A Shiny application object
}
\description{
A Shiny application for converting MSdial output to Jupyter format
}
\details{
Launch MSdial to Jupyter Converter App
}
\examples{
\dontrun{
run_msdial_to_jupyter()
}
}
