% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runMSdialApp}
\alias{runMSdialApp}
\title{MSdial to Jupyter Converter}
\usage{
runMSdialApp()
}
\value{
A Shiny application object
}
\description{
A Shiny application for converting MSdial output to Jupyter format
}
\details{
Launch MSdial to Jupyter Converter App
}
\examples{
\dontrun{
runMSdialApp()
}
}
