% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Score_IPF_Present}
\alias{Score_IPF_Present}
\title{Check is SCORE_IPF is in database}
\usage{
Score_IPF_Present(oswFile)
}
\arguments{
\item{oswFile}{A character vector pointing to path of osw file}
}
\value{
returns a logical value if SCORE_IPF is present or not
}
\description{
Check is SCORE_IPF is in database
}
