% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{set_continuous_axes}
\alias{set_continuous_axes}
\title{Defines ggplot panel by ratio, size, unit size, center, and axis breaks (\code{+}-able)}
\usage{
set_continuous_axes(
  x.axis.limits = NULL,
  y.axis.limits = NULL,
  x.axis.breaks = NULL,
  y.axis.breaks = NULL,
  aspect.ratio = 1,
  plot.center = NULL,
  axis.unit.ratio = NULL,
  coord_fun = coord_fixed,
  expand.x.axis = c(0, 0),
  expand.y.axis = c(0, 0)
)
}
\arguments{
\item{x.axis.limits}{vector containing lower and upper x-axis limits}

\item{y.axis.limits}{vector containing lower and upper y-axis limits}

\item{x.axis.breaks}{distance between x-axis breaks}

\item{y.axis.breaks}{distance between y-axis breaks}

\item{aspect.ratio}{absolute length of x-axis/y-axis}

\item{plot.center}{vector for center of plot}

\item{axis.unit.ratio}{ratio between x- and y-axis units}

\item{coord_fun}{coord function to apply the unit ratio with}

\item{expand.x.axis}{expand x.axis (see scale_x_continuous)}

\item{expand.y.axis}{expand y.axis (see scale_y_continuous)}
}
\value{
an object to be added to a ggplot with \code{+}
}
\description{
Add to a ggplot with \code{+}, e.g. \code{p + set_continuous_axes(axis.unit.ratio = 1)}.
Reads the built plot limits, so add it last in the chain.
}
