% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{set_PCA_labs}
\alias{set_PCA_labs}
\title{Add PCA axis labels (with \% variance) to a ggplot (\code{+}-able)}
\usage{
set_PCA_labs(PCx = NULL, PCy = NULL, sdev = NULL, digits = 1)
}
\arguments{
\item{PCx, PCy}{principal component numbers; inferred from the plot mapping if omitted}

\item{sdev}{vector of PC standard deviations; computed from the plot data if omitted}

\item{digits}{digits for the rounded percentage}
}
\value{
an object to be added to a ggplot with \code{+}
}
\description{
Add to a ggplot with \code{+}, e.g. \code{p + set_PCA_labs()}.
}
