% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_helpers.R
\name{strsplit_keep_first}
\alias{strsplit_keep_first}
\title{Performs strsplit and keeps first element of each string}
\usage{
strsplit_keep_first(x, split, as_numeric = F)
}
\arguments{
\item{x}{vector}

\item{split}{separator}

\item{as_numeric}{try to convert to numeric}
}
\description{
Performs strsplit and keeps first element of each string
}
