% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_helpers.R
\name{strsplit_keep_firstn}
\alias{strsplit_keep_firstn}
\title{Performs strsplit and keeps first n elements of each string}
\usage{
strsplit_keep_firstn(x, split, n = 1, as_numeric = F)
}
\arguments{
\item{x}{vector}

\item{split}{separator}

\item{n}{number of elements to keep counting from the beginning}

\item{as_numeric}{try to convert to numeric}
}
\description{
Performs strsplit and keeps first n elements of each string
}
