% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_helpers.R
\name{strsplit_keep_last}
\alias{strsplit_keep_last}
\title{Performs strsplit and keeps first n elements of each string}
\usage{
strsplit_keep_last(x, split, as_numeric = F)
}
\arguments{
\item{x}{vector}

\item{split}{separator}

\item{as_numeric}{try to convert to numeric}
}
\description{
Performs strsplit and keeps first n elements of each string
}
