% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{tiff_open}
\alias{tiff_open}
\title{Title}
\usage{
tiff_open(file)
}
\arguments{
\item{file}{tiff file to open (opens file saved under .tiff_temp if no argument given)}
}
\description{
Title
}
\examples{
tiff_temp()
plot(6, 9)
tiff_open
}
