% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCoord2Isocoord.R
\name{toolCoord2Isocoord}
\alias{toolCoord2Isocoord}
\title{toolCoord2Isocoord}
\usage{
toolCoord2Isocoord(x)
}
\arguments{
\item{x}{magclass object to be transformed from coordinates to iso-coordinate}
}
\value{
magpie object with 67420 cells in x.y.iso naming
}
\description{
Transforms an object with coordinate spatial data (on half-degree)
             to object with 67420 cells and coordinate and iso country information
}
\author{
Felicitas Beier, Michael Crawford
}
