% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolExpectLessDiff.R
\name{toolExpectLessDiff}
\alias{toolExpectLessDiff}
\title{toolExpectLessDiff}
\usage{
toolExpectLessDiff(x, y, maxdiff, description, maxdiff2 = NULL)
}
\arguments{
\item{x}{object 1}

\item{y}{object 2 which has the same format as object 1}

\item{maxdiff}{allowed maximum difference per element between x and y.}

\item{description}{a description of the check}

\item{maxdiff2}{optional additional threshold. If set it will serve as a second, critial threshold
which will throw a warning (instead of a simple note in case of \code{maxdiff}) if being surpassed.}
}
\description{
tool function for status reporting. It performs a difference check between
two objects and returns either a message via \code{toolStatusMessage}, that
the test was successful or that it failed.
}
\examples{
toolExpectLessDiff(1:3, 2:4, 10, "data is sufficiently close")
getMadratMessage("status")
}
\seealso{
\code{\link[madrat]{getMadratMessage}}, \code{\link{toolExpectTrue}}, \code{\link{toolStatusMessage}}
}
\author{
Jan Philipp Dietrich
}
