% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFertilizerDistribution.R
\name{toolFertilizerDistribution}
\alias{toolFertilizerDistribution}
\title{toolFertilizerDistribution}
\usage{
toolFertilizerDistribution(
  iterMax = 50,
  maxSnupe = 0.85,
  fertilizer,
  mapping = NULL,
  from = NULL,
  to = NULL,
  snupe,
  withdrawals,
  organicinputs,
  threshold = 0.5,
  progress = TRUE
)
}
\arguments{
\item{iterMax}{Maximum number of iterations for downscaling. 10 iterations are done at least.}

\item{maxSnupe}{Maximum allowed snupe (used to cap NA's during iteration)}

\item{fertilizer}{Total inorganic fertilizer to distribute (regional level)}

\item{mapping}{mapping used for disaggregation}

\item{from}{name of from column in mapping}

\item{to}{name of to column in mapping}

\item{snupe}{Regional soil nitrogen uptake efficiency}

\item{withdrawals}{Nitrogen withdrawals at the cell level}

\item{organicinputs}{Non-inorganic fertilizer inputs at the cell level}

\item{threshold}{Tg N difference below which we consider convergence "good enough"}

\item{progress}{Logical, should progress messages be printed}
}
\value{
magpie object of fertilizer usage at cell level
}
\description{
Disaggregates fertilizer usage to best match soil nitrogen uptake efficiency (snupe).
}
\author{
Benjamin Leon Bodirsky, Michael Crawford
}
