% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHoldConstant.R
\name{toolHoldConstant}
\alias{toolHoldConstant}
\title{toolHoldConstant}
\usage{
toolHoldConstant(x, years)
}
\arguments{
\item{x}{MAgPIE object to be continued.}

\item{years}{years for which the data should exist (hold constant, if missing)}
}
\value{
MAgPIE object with completed time dimensionality.
}
\description{
Holds a historical dataset constant for the entire period \code{years}.
}
\author{
Benjamin Leon Bodirsky, Jan Philipp Dietrich
}
