% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolWriteMadratLog.R
\name{toolWriteMadratLog}
\alias{toolWriteMadratLog}
\title{toolWriteMadratLog}
\usage{
toolWriteMadratLog(
  checkResults = getMadratMessage("status"),
  logPath = "status.log"
)
}
\arguments{
\item{checkResults}{list of check results as returned by \code{\link[madrat]{getMadratMessage}}}

\item{logPath}{path to the log file to be written}
}
\description{
Tool function for writing madrat messages to a log file. Useful
after running madrat calculations which are performing checks via
\code{\link{toolExpectTrue}} or other toolExpect functions.
}
\author{
Pascal Sauer
}
