% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimod_mapping-data.R
\docType{data}
\name{unimod_mapping}
\alias{unimod_mapping}
\title{A unimod_mapping table}
\format{An object of class "tbl_df"     "tbl"        "data.frame"}
\source{
\href{http://www.unimod.org/downloads.html}{UniMod}
}
\usage{
data(unimod_mapping)
}
\description{
A table containing mapping for UniMod ID accession and
modification code name.
}
\examples{
data(unimod_mapping)
}
\keyword{datasets}
