% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnimodMapping.R
\name{unimodTocodename}
\alias{unimodTocodename}
\title{Conversion from UniMod ID to modification name}
\usage{
unimodTocodename( mod_seq, out='sequence' )
}
\arguments{
\item{mod_seq}{A character vector containing of a peptide sequence, containing UniMod IDs to be converted to code names}

\item{out}{A character vector. Default: 'sequence' will return mod_seq with covnerted unimod to codename. Otherwise, will return converted codename}
}
\value{
A character vector. Returned mod_seq with converted codename or returned codename
}
\description{
Conversion from UniMod ID to modification name
}
\examples{
mod_seq <- "EGHAQNPM(UniMod:35)EPS(UniMod:21)VPQLS(UniMod:21)LM(UniMod:35)DVK"
## Will return peptide sequence with modifciation code name
unimodTocodename( mod_seq )
## Out
## [1] "EGHAQNPM(Oxidation)EPS(Phospho)VPQLS(Phospho)LM(Oxidation)DVK"
}
