% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnimodMapping.R
\name{unimodTouniisoid}
\alias{unimodTouniisoid}
\title{Conversion from UniMod ID to a unique string isomeric id}
\usage{
unimodTouniisoid( mod_seq, out='sequence' )
}
\arguments{
\item{mod_seq}{A character vector containing of a peptide sequence, containing UniMod IDs to be converted to code names}

\item{out}{A character vector. Default: 'sequence' will return mod_seq with covnerted unimod to codename. Otherwise, will return converted codename}

\item{mod_exclusion}{A vector of unimod record id modifications to exclude, and replace by nothing. Example: c("35")}

\item{mod_grouping}{A dataframe contained uninod ids and random characters to group isomers by. Example: data.frame( record_id=c(4, 259, 267), rand_char_id="B", stringsAsFactors = F )}
}
\value{
A character vector. Returned mod_seq with converted codename or returned codename
}
\description{
Conversion from UniMod ID to a unique string isomeric id
}
\examples{
mod_seq <- "EGHAQNPM(UniMod:35)EPS(UniMod:21)VPQLS(UniMod:21)LM(UniMod:35)DVK"
## Will return a unique string id that can be used to group isomers
unimodTouniisoid( mod_seq )
## Out
## [1] "EGHAQNPM(Oxidation)EPS(Phospho)VPQLS(Phospho)LM(Oxidation)DVK"
}
