% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSWData.R
\name{getOSWData_}
\alias{getOSWData_}
\title{Extract data from OpenSwathWorkflow .osw results file}
\usage{
getOSWData_(
  oswfile,
  run_name = "",
  precursor_id = "",
  peptide_id = "",
  peptide_unmodified = "",
  peptide_modified = "",
  mod_peptide_id = c("", ""),
  mod_residue_position = "",
  peak_group_rank_filter = FALSE,
  report_top_single_result = FALSE,
  pep_list = "",
  mscore_filter = c(SCORE_ = 1),
  ipf_filter = "",
  ipf_score = FALSE,
  ms2_score = TRUE,
  decoy_filter = TRUE,
  inference_level = "peptide_query",
  mod_exclusion = NULL,
  mod_grouping = NULL,
  remove_KR_label = NULL
)
}
\arguments{
\item{oswfile}{A character vector of the absolute path and filename of the osw results file. (Must be .osw format)}

\item{run_name}{A character vector for extraction of a specific run, this should be the same as the file name in the .OSW RUN table. (i.e. run_name='yanliu_I170114_016_PhosNoco4_SW.mzXML.gz', Default: '')}

\item{precursor_id}{A numeric value for a specific precursor id to extract information for. (Default: '')}

\item{peptide_id}{A numeric value for a specific peptide id to extract information for. (Default: '')}

\item{peptide_unmodified}{A string vector indicating a specific unmodified peptide sequence to extract information for. (Default: '')}

\item{peptide_modified}{A string vector indicating a specific modified peptide sequence to extract information for. (Default: '')}

\item{mod_peptide_id}{An array of two string vectors indicating a specific modified peptide sequence with both UniMod annotation and actual modification name to extract information for. I.e. c(ANS(Phos)SNSLK, ANS(UniMod:21)SNSLK) (Default: '')}

\item{mod_residue_position}{A numeric value indicating the position of a modification. (Default: '')}

\item{peak_group_rank_filter}{A Logical value for filtering OSW results by Peak-Group Rank of 1. (Default: FALSE)}

\item{report_top_single_result}{A logical value indificating to keep only the results with the lowest m_score. (Default: TRUE)}

\item{pep_list}{An arrary of string modified peptide sequences to extract information for. (Default: '')}

\item{mscore_filter}{A named numeric value to filter results by named q-value. (Default: '') Options: SCORE_MS2, SCORE_IPF, SCORE_PEPTIDE. Example: c(SCORE_MS2=0.01)}

\item{ipf_filter}{A numeric value to filter results by IPF PEP. (Default: '')}

\item{ipf_score}{A logical value to extract data using IPF Score results. (Default: FALSE)}

\item{ms2_score}{A logical value to extract data using MS2 Score results. (Default: TRUE)}

\item{decoy_filter}{A logical value to filter decoys out of final results. (Default: TRUE)}

\item{mod_exclusion}{A vector of unimod record id modifications to exclude, and replace by nothing. Example: c("35")}

\item{mod_grouping}{A dataframe contained uninod ids and random characters to group isomers by. Example: data.frame( record_id=c(4, 259, 267), rand_char_id="B", stringsAsFactors = F )}

\item{remove_KR_label}{A vector of UniMod Ids to remove from Arginine or Lysine. Example:  c("UniMod:259", "UniMod:267")}
}
\value{
A data.table containing OpenSwath Results information.
}
\description{
This function can be used to extract information from the OSW results file obtained from
running OpenSwathWorkflow
}
\author{
Justin Sing \url{https://github.com/singjc}
}
