% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransitionScores.R
\name{getTransitionScores_}
\alias{getTransitionScores_}
\title{Extract Transition Information from .osw results file}
\usage{
getTransitionScores_(oswfile, run_name, precursor_id = "", peptide_id = "")
}
\arguments{
\item{oswfile}{A character vector of the absolute path and filename of the osw results file. (Must be .osw format)}

\item{run_name}{A character vector for extraction of a specific run, this should be the same name as the file name in the .OSW RUN table.}

\item{precursor_id}{A numeric value for a specific precursor id to extract information for. (Default: '')}

\item{peptide_id}{A string vector indicatig a specific peptide to extract information for. (Default: '')}
}
\value{
A data.table containing Transition IDs and Transition Posterior Error Probabilities, and other useful information. Similar to getOSWData.R.
}
\description{
This function can be used to extract Transition Id and Transition Posterior Error Probability information from the OSW results file obtained from
running OpenSwathWorkflow
}
\author{
Justin Sing \url{https://github.com/singjc}
}
