% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_helpers.R
\name{length_unique_min2}
\alias{length_unique_min2}
\title{Count unique elemenmts with n >= 2}
\usage{
length_unique_min2(x)
}
\arguments{
\item{x}{}
}
\description{
Count unique elemenmts with n >= 2
}
