% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ordering.R
\name{num2idchar}
\alias{num2idchar}
\title{Transforms id numbers into cahracters with equal number of leading zero's}
\usage{
num2idchar(x, as.numeric = F)
}
\arguments{
\item{as.numeric}{}
}
\description{
Transforms id numbers into cahracters with equal number of leading zero's
}
