% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ordering.R
\name{partition_n2char}
\alias{partition_n2char}
\title{Transform values of x (row numbers) into groups of n as characters with equal
number of leading zero's}
\usage{
partition_n2char(x, n)
}
\arguments{
\item{n}{}
}
\description{
Transform values of x (row numbers) into groups of n as characters with equal
number of leading zero's
}
