% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolExpectLessDiff.R
\name{toolExpectLessDiff}
\alias{toolExpectLessDiff}
\title{toolExpectLessDiff}
\usage{
toolExpectLessDiff(x, y, maxdiff, description, level = 0)
}
\arguments{
\item{x}{object 1}

\item{y}{object 2 which has the same format as object 1}

\item{maxdiff}{allowed maximum difference per element between x and y.}

\item{description}{a description of the check}

\item{level}{as the test result will be linked to a function call, the function
needs to know to which call it should be linked. by default (\code{level = 0}) the parent
function call is being used. Increasing the number by one will let the function go
up by one in the call stack, \code{level = -1} will use \code{toolExpectTrue} itself as
function call.}
}
\description{
tool function for status reporting. It performs a difference check between
two objects and returns either a message via \code{toolStatusMessage}, that
the test was successful or that it failed.
}
\examples{
toolExpectLessDiff(1:3, 2:4, 10, "data is sufficiently close", level = -1)
getMadratMessage("status")
}
\seealso{
\code{\link{getMadratMessage}}, \code{\link{toolExpectTrue}}, \code{\link{toolStatusMessage}}
}
\author{
Jan Philipp Dietrich
}
