% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFertilizerDistribution.R
\name{toolFertilizerDistribution}
\alias{toolFertilizerDistribution}
\title{toolFertilizerDistribution}
\usage{
toolFertilizerDistribution(
  iterMax = 50,
  maxSnupe = 0.85,
  fertilizer,
  snupe,
  withdrawals,
  organicinputs,
  threshold = 0.5
)
}
\arguments{
\item{iterMax}{Maximum number of iterations for downscaling}

\item{maxSnupe}{Maximum allowed snupe (used to cap NA's during iteration)}

\item{fertilizer}{Total inorganic fertilizer to distribute (regional level)}

\item{snupe}{Regional soil nitrogen uptake efficiency}

\item{withdrawals}{Nitrogen withdrawals at the cell level}

\item{organicinputs}{Non-inorganic fertilizer inputs at the cell level}

\item{threshold}{Tg N difference below which we consider convergence "good enough"}
}
\value{
magpie object of fertilizer usage at cell level
}
\description{
Disaggregates fertilizer usage to best match soil nitrogen uptake efficiency (snupe).
}
