% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFertilizerDistribution.R
\name{toolFertilizerDistribution}
\alias{toolFertilizerDistribution}
\title{toolFertilizerDistribution}
\usage{
toolFertilizerDistribution(
  iteration_max = 50,
  max_snupe = 0.85,
  mapping,
  from,
  to,
  fertilizer,
  SNUpE,
  withdrawals,
  organicinputs,
  threshold = 0.5
)
}
\arguments{
\item{iteration_max}{maximum iteration for downscaling}

\item{max_snupe}{the maximum level of nue or snupe}

\item{mapping}{mapping used for disaggregation}

\item{from}{name of from column in mapping}

\item{to}{name of to column in mapping}

\item{fertilizer}{total inorganic fertilizer to be distributed on regional leve}

\item{SNUpE}{Nitrogen use efficiency or SNUPE on regional level which should be matched best possible}

\item{withdrawals}{nitrogen withdrawals on cell level}

\item{organicinputs}{non-inroganic fertilizer inputs on cell level}

\item{threshold}{threshold in Tg Nr until when the distribution counts as converged}
}
\value{
magpie object with fertilizer usage on cell level
}
\description{
Disaggregates fertilizer usage, trying to best match a certain soil nitrogen uptake efficiency (SNUpE). Also used in magpie4 library
}
\author{
Benjamin Leon Bodirsky
}
