% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolIso2CellCountries.R
\name{toolIso2CellCountries}
\alias{toolIso2CellCountries}
\title{toolIso2CellCountries}
\usage{
toolIso2CellCountries(x, cells = "lpjcell", absolute = NULL)
}
\arguments{
\item{x}{magpie object on iso country level}

\item{cells}{switch between 59199 ("magpiecell") and 67420 ("lpjcell") cells}

\item{absolute}{switch declaring the values as absolute (TRUE) or relative (FALSE)
for additional (type-specific) diagnostic information. If not defined (NULL) additional diagnostics
will not be shown.}
}
\value{
return selected input data
}
\description{
Select country names of countries which are present on cellular level
}
\author{
Kristine Karstens, Felicitas Beier, Jan Philipp Dietrich
}
