% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XICMasterPlotFcn.R
\name{XICMasterPlotFcn_}
\alias{XICMasterPlotFcn_}
\title{Master Plotting function for plotting XIC's}
\usage{
XICMasterPlotFcn_(
  dup_peps,
  uni_mod = NULL,
  sqMass_files,
  in_lib,
  in_osw,
  plotPrecursor = T,
  plotIntersectingDetecting = T,
  plotUniqueDetecting = T,
  plotIdentifying = T,
  plotIdentifying.Unique = T,
  plotIdentifying.Shared = F,
  plotIdentifying.Against = F,
  smooth_chromatogram = list(p = 4, n = 9),
  doFacetZoom = T,
  FacetFcnCall = NULL,
  doPlot = T,
  RT_padding = 100,
  Charge_State = NULL,
  N_sample = 2,
  idx_draw_these = NULL,
  store_plots_subdir = "/Results/",
  printPlot = F,
  use_top_trans_pep = F,
  show_n_transitions = NULL,
  show_all_pkgrprnk = T,
  show_legend = T,
  verbosity = 0
)
}
\arguments{
\item{dup_peps}{A character vector of a peptide sequence}
}
\value{
A ggplot-grobs table of a XIC
}
\description{
This function can be used to draw XIC's by calling getXIC
}
\author{
Justin Sing \url{https://github.com/singjc}
}
